refactor: move presentation components to `ui_core` package
