+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.WeatherAppDemo">
         <activity
-            android:name=".presentation.MainActivity"
+            android:name=".presentation.ui_core.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.WeatherAppDemo">
             <intent-filter>
