+++ app/src/main/java/com/axehai/weatherscope/presentation/ui_core/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.axehai.weatherscope.presentation
+package com.axehai.weatherscope.presentation.ui_core
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -11,7 +11,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.axehai.weatherscope.presentation.ui.theme.WeatherAppDemoTheme
+import com.axehai.weatherscope.presentation.ui_core.theme.WeatherAppDemoTheme
 
 class MainActivity : ComponentActivity() {
 	override fun onCreate(savedInstanceState: Bundle?) {
