+++ app/src/main/java/com/axehai/weatherscope/presentation/ui_core/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.axehai.weatherscope.presentation.ui.theme
+package com.axehai.weatherscope.presentation.ui_core.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
